+++ module/Batch/src/Command/FaucetClaimStats/FaucetClaimStatsFactory.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Batch\Command\FaucetClaimStats;
+
+class FaucetClaimStatsFactory {
+    public function __invoke($controllers): FaucetClaimStats
+    {
+        return new FaucetClaimStats($controllers->get('faucetdev'));
+    }
+}