+++ module/Batch/src/Tools/BatchTools.php
@@ -2,16 +2,37 @@
 
 namespace Batch\Tools;
 
+use Laminas\Db\TableGateway\TableGateway;
+use Laminas\Db\Sql\Where;
+
 class BatchTools {
+    /**
+     * Guild Weekly Task Status Table
+     *
+     * @var TableGateway $mGuildTaskStatusTbl
+     * @since 1.0.0
+     */
+    protected TableGateway $mGuildTaskStatusTbl;
+
+    /**
+     * Guild User Table
+     *
+     * @var TableGateway $mGuildUserTbl
+     * @since 1.0.0
+     */
+    protected TableGateway $mGuildUserTbl;
+
     /**
      * Constructor
      *
      * BatchTools constructor.
      * @since 1.0.0
      */
-    public function __construct()
+    public function __construct($mapper)
     {
-
+        // Guild Weekly Task Addon
+        $this->mGuildUserTbl = new TableGateway('faucet_guild_user', $mapper);
+        $this->mGuildTaskStatusTbl = new TableGateway('faucet_guild_weekly_status', $mapper);
     }
 
     /**
@@ -58,5 +79,80 @@ class BatchTools {
         }
 
         return $week;
+    }
+
+    /**
+     * Load All Users in an active Guild
+     * @return array
+     */
+    public function loadUsersInGuilds(): array
+    {
+        $guWh = new Where();
+        $guWh->notLike('date_joined', '0000-00-00 00:00:00');
+        $guildUsers = $this->mGuildUserTbl->select($guWh);
+        $guildByUserId = [];
+        foreach ($guildUsers as $gu) {
+            $guildByUserId['user-' . $gu->user_idfs] = $gu->guild_idfs;
+        }
+
+        return $guildByUserId;
+    }
+
+    /**
+     * Update Guild Weekly Stats
+     * @param $stats
+     * @param $week
+     * @param $statKey
+     */
+    public function updateGuildWeeklyStats($stats, $week, $statKey, $output) : bool
+    {
+
+        $weekInfo = explode('-', $week);
+        $weekNo = $weekInfo[0];
+        $yearNo = $weekInfo[1];
+        $guildCount = 0;
+        if(count($stats) == 0) {
+            $output->writeln([
+                '-- No updates for Guild Weekly Tasks ('.$statKey.') for Week '.$week,
+            ]);
+            return true;
+        }
+        $output->writeln([
+            '-- Update Guild Weekly Stats ('.$statKey.') for Week '.$week,
+        ]);
+        foreach(array_keys($stats) as $guildKey) {
+            $guildId = substr($guildKey, strlen('guild-'));
+            if(is_numeric($guildId)) {
+                $guildCount++;
+                $check = $this->mGuildTaskStatusTbl->select([
+                    'weekly_key' => $statKey,
+                    'guild_idfs' => $guildId,
+                    'week' => $weekNo,
+                    'year' => $yearNo
+                ]);
+                if($check->count() == 0) {
+                    $this->mGuildTaskStatusTbl->insert([
+                        'weekly_key' => $statKey,
+                        'guild_idfs' => $guildId,
+                        'week' => $weekNo,
+                        'year' => $yearNo,
+                        'progress' => $stats[$guildKey]
+                    ]);
+                } else {
+                    $checkInfo = $check->current();
+                    $statId = $checkInfo->id;
+                    $currentVal = $checkInfo->progress;
+                    $this->mGuildTaskStatusTbl->update([
+                        'progress' => $currentVal+$stats[$guildKey]
+                    ],['id' => $statId]);
+                }
+            }
+        }
+
+        $output->writeln([
+            '-- Updated Stats for '.$guildCount.' Guilds',
+        ]);
+
+        return true;
     }
 }