+++ CHANGELOG.md
@@ -5,9 +5,18 @@ All notable changes to this project will be documented in this file, in reverse 
 ## Todo
 - Add "hacker" achievement to MinerShares
 
+
+## 1.2.0
+
+### Added
+- Check PTC Crypto Payments Batch
+- Offerwall Stats Batch
+- Faucet Claim Stats Batch
+
 ## 1.1.0
 
 ### Added
+- Daily Tasks Batch
 - Miner Payments Batch
 - Miner Shares Batch
 - Miner Stats Batch